[#44] Refactor code
